feat: 自定义机场噪音等级按照格数计算而非按照跑道数量计算

+++ src/script/api/script_airport.cpp
@@ -138,7 +138,7 @@
 
 extern CommandCost AddAirportTileTableToBitmapTileArea(const AirportTileLayout &atl, BitmapTileArea *bta, DiagDirection rotation, uint cost_multiplier);
 extern Town *AirportGetNearestTown(BitmapTileArea bta, uint &mindist);
-extern uint8_t GetAirportNoiseLevelForDistance(uint noise_level, uint distance);
+extern uint16_t GetAirportNoiseLevelForDistance(uint noise_level, uint distance);
 
 /* static */ SQInteger ScriptAirport::GetNoiseLevelIncrease(TileIndex tile, AirportType type)
 {
